requirements.txt
appdirs==1.4.3
click==6.7
dj-database-url==0.4.2
Django==1.11.1
django-ckeditor==5.2.2
django-tagging==0.4.5
gunicorn==19.7.1
olefile==0.44
packaging==16.8
Pillow==4.1.1
pyparsing==2.2.0
pytz==2017.2
requests==2.14.2
six==1.10.0
